refactor: Remove duplicate type declarations.
