+++ back/src/types/index.ts
@@ -45,3 +45,12 @@ export interface GetCommitDetailListRequest {
   repoName: string;
   commitHash: string;
 }
+
+export interface CommitFile {
+  filename: string;
+  status: "added" | "modified" | "deleted" | "renamed";
+  additions: number;
+  deletions: number;
+  changes: number;
+  patch: string;
+}
